backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
yfinance==0.2.37
python-dotenv==1.0.1
feedparser==6.0.10
requests==2.31.0
slowapi==0.1.9
